Fixed a few warnings.
Rev: src/builtin_functions.c:1.299
Rev: src/modules/HTTPLoop/cache.c:1.14
Rev: src/modules/HTTPLoop/cache.h:1.7
Rev: src/modules/HTTPLoop/requestobject.c:1.14
Rev: src/modules/Image/encodings/bmp.c:1.28
Rev: src/peep.c:1.36
Rev: src/program.c:1.264

+++ src/modules/HTTPLoop/cache.c
@@ -168,7 +168,7 @@ static void really_free_cache_entry(struct cache  *c, struct cache_entry *e,
 
 
 void aap_free_cache_entry(struct cache *c, struct cache_entry *e,
-			  struct cache_entry *prev, int b)
+			  struct cache_entry *prev, size_t b)
 {
 #ifdef DEBUG
   if(e->refs<=0)
@@ -205,7 +205,7 @@ void aap_cache_insert(struct cache_entry *ce, struct cache *c)
 {
   struct cache_entry *head, *p;
   char *t;
-  int hv;
+  size_t hv;
 #ifdef DEBUG
   extern int d_flag;
   if((d_flag > 2) && !mt_trylock( & c->mutex ))
